+++ src/features/basic/index.ts
@@ -35,6 +35,7 @@ import './lm-highlight-own-ads';
 import './lm-item-icons/lm-item-icons';
 import './mat-linkify-category';
 import './mtra-auto-focus-amount';
+import './mtra-transfer-on-enter';
 import './nots-material-ticker';
 import './nots-notification-type-label';
 import './nots-ship-arrival-inventory';