+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - `mtra-auto-focus-amount`: `MTRA`: Automatically focuses the amount input on buffer open
+- `mtra-transfer-on-enter`: `MTRA`: Triggers transfer on Enter and closes the buffer on success
 
 ## 25.1.28